# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Apiodactyl API",
    "version": "1.0.0",
    "description": "API specification for games, books, reviews, projects, screenshots, and admin endpoints."
  },
  "servers": [
    {
      "url": "https://api.bearodactyl.dev"
    },
    {
      "url": "http://127.0.0.1:2379"
    }
  ],
  "paths": {
...
    "/read-watch/{book_id}": {
      "get": {
        "summary": "Get book by ID",
        "operationId": "getBookById",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Book details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update book",
        "operationId": "updateBook",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewBook"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch book",
        "operationId": "patchBook",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBook"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Book patched",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete book",
        "operationId": "deleteBook",
        "parameters": [
          {
            "name": "book_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Book deleted"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string",
              "nullable": true
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "nullable": true
            }
          },
          "rating": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "my_thoughts": {
            "type": "string"
          },
          "links": {
            "type": "array",
            "items": {
              "type": "object",
              "nullable": true
            },
            "nullable": true
          },
          "cover_image": {
            "type": "string"
          },
          "explicit": {
            "type": "boolean"
          },
          "color": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "author",
          "genres",
          "tags",
          "rating",
          "status",
          "description",
          "my_thoughts",
          "cover_image",
          "explicit"
        ]
      },
      "NewBook": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rating": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "my_thoughts": {
            "type": "string"
          },
          "links": {
            "type": "array",
            "items": {
              "type": "object",
              "nullable": true
            },
            "nullable": true
          },
          "cover_image": {
            "type": "string"
          },
          "explicit": {
            "type": "boolean"
          },
          "color": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "title",
          "author",
          "genres",
          "tags",
          "rating",
          "status",
          "description",
          "my_thoughts",
          "cover_image",
          "explicit"
        ]
      },
      "UpdateBook": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rating": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "my_thoughts": {
            "type": "string"
          },
          "links": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "nullable": true
          },
          "cover_image": {
            "type": "string"
          },
          "explicit": {
            "type": "boolean"
          },
          "color": {
            "type": "string"
          }
        }
      },
...
    },
...
  }
}